pyproject.toml
[tool.poetry]
name = "lmnop"
version = "0.0.0"  # DO NOT CHANGE -- set during build
description = "Build and deploy functions."
authors = ["Paul Lockaby <paul@paullockaby.com>"]
packages = [{include = "lmnop", from = "src"}]

[tool.poetry.scripts]
builder = "lmnop.builder.app:run"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.1"
requests = "^2.26.0"
invoke = "^1.6.0"
PyYAML = "^6.0"
schema = "^0.7.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-flake8 = "^1.0.7"
flake8 = "^3.9.2"
flake8-bandit = "^2.1.2"
flake8-builtins = "^1.5.3"
flake8-bugbear = "^21.9.2"
flake8-comprehensions = "^3.7.0"
flake8-isort = "^4.0.0"
pep8-naming = "^0.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--flake8 -p no:warnings --log-level=ERROR"
